requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
Cerberus==1.3.4
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.2.0
click==7.1.2
deep-translator==1.11.4
ecdsa==0.13
fillpdf==0.7.2
Flask==1.1.4
frozenlist==1.4.0
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
magic-filter==1.0.10
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector==2.2.9
numpy==1.26.4
pandas==2.2.2
pdf2image==1.16.3
pdfrw2==0.5.0
Pillow==10.0.0
polib==1.2.0
py==1.11.0
pycryptodome==3.18.0
PyJWT==1.7.1
PyMuPDF==1.22.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
sqlcollection==0.3.0
tzdata==2024.1
urllib3==2.0.5
user-api==0.6.0
Werkzeug==1.0.1
yarl==1.9.2
